third_party/blink/web_tests/platform/generic/cssom/ahem-ex-units-expected.txt

Test ex unit with Ahem font

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS parseFloat(computedStyleDiv("word-spacing", "1ex")) is within 0.05 of 12.8
PASS parseFloat(computedStyleRect("word-spacing", "1ex")) is within 0.05 of 12.8
PASS successfullyParsed is true

TEST COMPLETE
